Moved things into the Asset Test Scene. NPC templating & a Yarn tutorial!
To make a new NPC, go to Assets/Prefabs/NPCs/ and copy and paste the NPC_Template prefab and double click your new NPC to edit it.
On the main object, change 'My Dialogue Label' in the inspector to match whatever label's in the matching Yarn script for the new character!
Then go down the tree in Hierarchy to "CharacterModel", and in the top of the inspector change the Prefab to be whatever model you want it to be. Animations might be tricky... but we'll figure that part out as we go.

For the Yarn tutorial, talk to the Rose NPC in the Asset Test Scene, and open her Yarn script in Assets/Dialogue/ to follow along!